Divid Components at service page
